Partial work towards modularization.
