+++ smvp-algs/CMakeLists.txt
@@ -1,2 +1,1 @@
-add_library(smvp-csr smvp-csr.c)
-add_library(smvp-tjds smvp-tjds.c)
+add_library(smvp-algs smvp-algs.c)
