--- smvp-algs/smvp-csr.c
@@ -1,6 +1,0 @@
-/* 
-*  ==================================================================
-*  smvp-csr.c v0.4 for smvp-toolbox
-*  ==================================================================
-*/
-
