--- smvp-algs/smvp-csr.h
@@ -1,13 +1,0 @@
-/* 
-*  ==================================================================
-*  smvp-csr.h v0.4 for smvp-toolbox
-*  ==================================================================
-*/
-
-
-#ifndef SMVP_CSR
-#define SMVP_CSR
-
-
-
-#endif
