--- smvp-algs/smvp-tjds.h
@@ -1,14 +1,0 @@
-/* 
-*   Matrix Market I/O library for ANSI C
-*
-*   See http://math.nist.gov/MatrixMarket for details.
-*
-*
-*/
-
-#ifndef MATH_SPMV
-#define MATH_SPMV
-
-
-
-#endif
